# Remove unnecessary parameter table from pulse program