--- a/qiskit/pulse/library/waveform.py
+++ b/qiskit/pulse/library/waveform.py
@@ -15,10 +15,8 @@ from typing import Dict, List, Optional, Union, Any
 
 import numpy as np
 
-from qiskit.circuit.parameterexpression import ParameterExpression, ParameterValueType
 from qiskit.pulse.exceptions import PulseError
 from qiskit.pulse.library.pulse import Pulse
-from qiskit.pulse.utils import deprecated_functionality
 
 
 class Waveform(Pulse):
@@ -103,13 +101,6 @@ class Waveform(Pulse):
     def parameters(self) -> Dict[str, Any]:
         """Return a dictionary containing the pulse's parameters."""
         return dict()
-
-    @deprecated_functionality
-    def assign_parameters(self,
-                          value_dict: Dict[ParameterExpression, ParameterValueType]
-                          ) -> 'Waveform':
-        # Waveforms don't accept parameters
-        return self
 
     def __eq__(self, other: Pulse) -> bool:
         return super().__eq__(other) and self.samples.shape == other.samples.shape and \