<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary
In #5854 parameter table is deprecated and `ParameterManager` is introduced. This PR removes unnecessary parameter tables and `.parameters`, `.assign_parameters` methods from pulse instructions and channels.


### Details and comments
Simple benchmark
```python
%%time

for _ in range(100):
    sched = pulse.ScheduleBlock()
    for _ in range(3000):
        sched += pulse.Play(pulse.Gaussian(circuit.Parameter('dur'), 
                                           circuit.Parameter('amp'), 
                                           circuit.Parameter('sigma')), 
                            pulse.DriveChannel(circuit.Parameter('ch')))
```
This PR:
CPU times: user 35.8 s, sys: 1.1 s, total: 36.9 s
Wall time: 36.9 s

master:
CPU times: user 46.2 s, sys: 1.36 s, total: 47.6 s
Wall time: 47.6 s

This improvement is simply thanks to the removal of parameter table. All instructions and channels had been initializing the table in the constructor, but this is just an overhead to create new instance. We just added deprecation warning in 0.17, perhaps this PR should be on hold. However, considering the performance improvement, it could be better to remove them soon.