@mtreinish 2021-04-07
Per the deprecation policy: https://qiskit.org/documentation/contributing_to_qiskit.html#deprecation-policy this is way too soon to remove. I'm marking this as on hold until we have given users a chance to adapt based on the deprecation.